pyproject.toml
[project]
name = "bandjacks"
version = "0.1.0"
description = "Cyber Threat Defense World Modeling System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "fuzzywuzzy>=0.18.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jsonschema>=4.0.0",
    "litellm>=1.0.0",
    "markdown>=3.8.2",
    "neo4j>=5.28.2",
    "opensearch-py>=3.0.0",
    "orjson>=3.11.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.10.1",
    "rdflib>=7.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "python-multipart>=0.0.6",
    "redis>=6.4.0",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "uvicorn[standard]>=0.35.0",
    "google-adk==1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-python-dateutil>=2.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
